+++ .github/workflows/Deploy_Service_Auth.yml
@@ -23,7 +23,8 @@ jobs:
           touch .env
           echo "VITE_SERVER_URL=${{ secrets.VITE_SERVER_URL }}" >> .env
           echo "VITE_KAKAO_RESTAPI_KEY=${{ secrets.VITE_KAKAO_RESTAPI_KEY }}" >> .env
-          echo "VITE_REDIRECT_URL=${{ secrets.VITE_REDIRECT_URL }}" >> .env
+          echo "VITE_HOST_URL=${{ secrets.VITE_HOST_URL }}" >> .env
+          echo "VITE_AUTH_URL=${{ secrets.VITE_AUTH_URL }}" >> .env
 
       - name: Get the version
         id: get_version
