+++ micro_services/auth/src/App.tsx
@@ -1,15 +1,14 @@
 import { Routes, Route, useLocation } from 'react-router-dom';
 
-import PageRoutes from 'auth/routes/page';
-import ComponentRoutes from 'auth/routes/component';
+import LoginPage from 'auth/pages/login';
 
 function App() {
   const location = useLocation();
 
   return (
     <Routes location={location}>
-      <Route element={<ComponentRoutes />} path="/components/*" />
-      <Route element={<PageRoutes />} path="/pages/*" />
+      <Route element={<LoginPage />} path="/login" />
+      <Route element={<>존재하지 않는 페이지 입니다</>} path="*" />
     </Routes>
   );
 }
