+++ micro_services/auth/src/pages/login/KaKaoLoginButton.tsx
@@ -1,10 +1,15 @@
+import { useState } from 'react';
+
 import Button from '@packages/components/Button';
 
 import { REQUEST_KAKAO_LOGIN_URL } from './constant';
 import kakaoIcon from './assets/kakaoIcon';
 
 export default function KaKaoLoginButton() {
+  const [loading, setLoading] = useState<boolean>(false);
+
   const handleKaKaoLogin = () => {
+    setLoading(true);
     window.open(REQUEST_KAKAO_LOGIN_URL, '_self', 'width=500,height=600');
   };
 
@@ -14,6 +19,7 @@ export default function KaKaoLoginButton() {
       customClass="bg-[#F7E600]"
       iconCustom={kakaoIcon}
       value={'카카오 계정으로 로그인'}
+      isLoading={loading}
       spinnerColor={'#212121'}
     />
   );
