+++ micro_services/auth/src/pages/login/assets/kakaoIcon.tsx
