+++ micro_services/auth/src/pages/login/constant.ts
@@ -1,0 +1,3 @@
+const { VITE_KAKAO_RESTAPI_KEY, VITE_AUTH_URL } = import.meta.env;
+
+export const REQUEST_KAKAO_LOGIN_URL = `https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=${VITE_KAKAO_RESTAPI_KEY}&redirect_uri=${VITE_AUTH_URL}/login`;
