--- micro_services/auth/src/routes/page.tsx
@@ -1,11 +1,0 @@
-import { Suspense } from 'react';
-
-import Loader from '@packages/components/Loader';
-
-export const PageRoutes = () => (
-  // const location = useLocation();
-
-  <Suspense fallback={<Loader />}></Suspense>
-);
-
-export default PageRoutes;
