+++ package.json
@@ -35,28 +35,33 @@
     "test": "jest --passWithNoTests",
     "generate:css": "npx tailwindcss -i ./packages/styles/_mobile.css -o ./packages/styles/tailwind-LEGACY.css",
     "prepare": "husky install",
-    "lint-staged": "npm run lint && lint-staged"
+    "lint-staged": "lint-staged"
   },
   "lint-staged": {
     "app/**/*.{js?(x),ts?(x)}": [
       "eslint --fix",
-      "prettier --write"
+      "prettier --write",
+      "bash -c 'npm run lint:main'"
     ],
      "packages/**/*.{js?(x),ts?(x)}": [
       "eslint --fix",
-      "prettier --write"
+      "prettier --write",
+      "bash -c 'npm run lint:packages'"
     ],
     "micro_services/auth/**/*.{js?(x),ts?(x)}": [
       "eslint --fix",
-      "prettier --write"
+      "prettier --write",
+      "bash -c 'npm run lint:auth'"
     ],
     "micro_services/books/**/*.{js?(x),ts?(x)}": [
       "eslint --fix",
-      "prettier --write"
+      "prettier --write",
+      "bash -c 'npm run lint:books'"
     ],
     "micro_services/host/**/*.{js?(x),ts?(x)}": [
       "eslint --fix",
-      "prettier --write"
+      "prettier --write",
+      "bash -c 'npm run lint:host'"
     ]
   },
   "workspaces": [
